+++ src/lib/gpu/webgpuOptimizer.ts
